requirements.txt
requests==2.32.3
discord.py==2.4.0
pytubefix==8.10.2
PyNaCl==1.5.0
psutil==5.9.4
RPi.GPIO==0.7.1
gpiozero==2.0.1
pillow==11.1.0
spidev==3.6
numpy==2.2.3